+++ gTile@shuairan/src/base/ui/AutoTileTwoList.ts
@@ -23,7 +23,7 @@ export class AutoTileTwoList extends ActionButton<"resize-done"> {
 
     this.app.platform.reset_window(this.app.FocusMetaWindow);
 
-    let monitor = this.app.Grid.monitor;
+    let monitor = this.app.CurrentMonitor;
     let [screenX, screenY, screenWidth, screenHeight] = getUsableScreenArea(monitor);
     let windows = this.app.GetNotFocusedWindowsOfMonitor(monitor);
     let nbWindowOnEachSide = Math.ceil((windows.length + 1) / 2);