+++ gTile@shuairan/test.sh
@@ -1,3 +1,19 @@
 #!/bin/bash
+
+SOURCE="${BASH_SOURCE[0]}"
+while [ -h "$SOURCE" ]; do # resolve $SOURCE until the file is no longer a symlink
+  DIR="$( cd -P "$( dirname "$SOURCE" )" >/dev/null && pwd )"
+  SOURCE="$(readlink "$SOURCE")"
+  [[ $SOURCE != /* ]] && SOURCE="$DIR/$SOURCE" # if $SOURCE was a relative symlink, we need to resolve it relative to the path where the symlink file was located
+done
+DIR="$( cd -P "$( dirname "$SOURCE" )" >/dev/null && pwd )"
+
+# Save current dir for convenience
+path=${PWD}
+
+cd $DIR
 cp -rf files/gTile@shuairan/* ~/.local/share/cinnamon/extensions/gTile@shuairan/
+cd ..
+./cinnamon-spices-makepot gTile@shuairan
+cd $PWD
 export DISPLAY=:0; cinnamon --replace &