+++ gTile@shuairan/CHANGELOG.md
@@ -1,3 +1,11 @@
+### 2.3.0
+
+* Add back old behavior as an option - Display UI on all monitors at the same time
+* Fixes [#421](https://github.com/linuxmint/cinnamon-spices-extensions/issues/421) Add back old behavior as an option - UI can have a fixed size instead of respecting the monitor's aspect ratio
+* Add back old behavior as an option - Always use monitor's center
+* Fix Auto-Tiling not working in 5.4 version 
+* Fixes [#420](https://github.com/linuxmint/cinnamon-spices-extensions/issues/420)
+
 ### 2.2.0
 
 * Split codebase after Cinnamon version 5.4 because of breaking changes and fix breakages.